+++ server/src/loaders/express.js
@@ -4,8 +4,6 @@ const { api } = require('../config');
 
 module.exports = app => {
   app.enable('trust proxy');
-
-  app.use(cors);
-
+  app.use(cors());
   app.use(api.prefix, routes());
 };
