+++ server/src/loaders/socket.js
@@ -1,9 +1,9 @@
 const handlerFactory = require('../socket/handlers');
-const handleSession = require('../socket/session');
+const sessionMiddleware = require('../socket/session-middleware');
 const handleClient = require('../socket/client');
 
 module.exports = io => {
   const makeHandlers = handlerFactory(io);
-  io.use(handleSession);
+  io.use(sessionMiddleware);
   io.on('connection', client => handleClient(client, makeHandlers(client)));
 };
