+++ server/src/socket/session-middleware.js
